Déplacement de floppy dans kernel/drivers, et ouverture différente selon block/char dans devfs

+++ kernel/include/fs/devfs.h
@@ -49,8 +49,8 @@ typedef struct {
 }chardev_interfaces;
 
 typedef struct {
-	size_t (*read)(open_file_descriptor *,void*, unsigned int, size_t);
-	size_t (*write)(open_file_descriptor *, const void*, unsigned int, size_t);
+	size_t (*read)(open_file_descriptor *,void*, size_t, unsigned int);
+	size_t (*write)(open_file_descriptor *, const void*, size_t, unsigned int);
 	int (*ioctl) (open_file_descriptor*, unsigned int, void*);
 	int (*open) (open_file_descriptor*);
 	int (*close) (open_file_descriptor*);
